app/requirements.txt
requests==2.31.0
pyExploitDb==0.2.9
python-dotenv==1.0.0